SGRIDDOC166 - For 11.9 Change max number of supported storage volume LUNs from 16 to 48 DEV: https://jira.ngage.netapp.com/browse/SGWS-31418 (closed) DOC: https://jira.ngage.netapp.com/browse/SGRIDDOC-166
In 11.9 we updated the limit for bare metal LUNs from 16 to 48. Original comment made against VMware, but info for all bare metal platforms appeared impacted and was changed:

https://review.docs.netapp.com/us-en/storagegrid_sgriddoc166-max-disk-count-kp-26june/expand/adding-storage-volumes-to-storage-nodes.html
https://review.docs.netapp.com/us-en/storagegrid_sgriddoc166-max-disk-count-kp-26june/expand/guidelines-for-adding-object-capacity.html
https://review.docs.netapp.com/us-en/storagegrid_sgriddoc166-max-disk-count-kp-26june/expand/linux-adding-direct-attached-or-san-volumes-to-storage-node.html
https://review.docs.netapp.com/us-en/storagegrid_sgriddoc166-max-disk-count-kp-26june/expand/vmware-adding-storage-volumes-to-storage-node.html
https://review.docs.netapp.com/us-en/storagegrid_sgriddoc166-max-disk-count-kp-26june/swnodes/storage-and-performance-requirements.html
There is a related issue for increased volume size that remains under dev discussion. Any volume size doc changes will be handled in a separate request.

This change is in the 12.0 consolidated bare metal topics. Like changes required for 11.9 in RHEL, Ubuntu/Debian, and VMware topics. These changes will be handled in a separate 11.9 PR once approved for 12.0.